fix: Kotlin equality between Int and Long is always false
NarrowLiteralInstruction.narrowLiteral returns Int, but the comparison
== 0L requires Long equality, which Kotlin evaluates by type first.
The result is never true, so the const/4 with value 0 after the
flag's key was never matched, and the patch aborted at patch time.

Widen with .toLong() before comparing.

+++ patches/src/main/kotlin/dev/jz6/flexboard/patches/features/grammarcheck/ForceGrammarCheckPatch.kt
@@ -82,7 +82,7 @@ private fun MutableMethod.forceDefaultOn() {
     val tail = instructions.drop(keyIndex + 1)
     val defaultIndex = keyIndex + 1 + tail.indexOfFirst { instruction ->
         instruction is OneRegisterInstruction && when (instruction) {
-            is NarrowLiteralInstruction -> instruction.narrowLiteral
+            is NarrowLiteralInstruction -> instruction.narrowLiteral.toLong()
             is WideLiteralInstruction -> instruction.wideLiteral
             else -> null
         } == 0L
